backend/requirements.txt

fastapi
uvicorn[standard]


motor
chromadb


google-generativeai

PyMuPDF
Pillow


python-dotenv
thefuzz


numpy
protobuf
pymongo

pymupdf4llm
python-multipart